+++ src/tuple.rs
@@ -67,6 +67,7 @@ fn imp_core_part(generics: &[Ident], indices: &[Index]) -> pm2::TokenStream {
             #[::core::prelude::v1::derive(
                 ::core::cmp::PartialEq, ::core::cmp::Eq, ::core::cmp::PartialOrd, ::core::cmp::Ord,
                 ::core::hash::Hash,
+                ::core::default::Default,
                 Clone, Copy,
             )]
             struct Tuple;
@@ -97,6 +98,7 @@ fn imp_core_part(generics: &[Ident], indices: &[Index]) -> pm2::TokenStream {
             #[::core::prelude::v1::derive(
                 ::core::cmp::PartialEq, ::core::cmp::Eq, ::core::cmp::PartialOrd, ::core::cmp::Ord,
                 ::core::hash::Hash,
+                ::core::default::Default,
                 Clone, Copy,
             )]
             struct Tuple<#(#generics),*>(#(#generics),*);